requirements.txt
pandas~=2.2.3
numpy~=2.2.4
click~=8.1.8
PyYAML~=6.0.2